+++ class-components/src/pages/index.ts
@@ -1,4 +1,4 @@
-import { HomePage } from './HomePage';
+import { HomePage } from './HomePage/HomePage';
 import { Page404 } from './Page404';
 
 export { HomePage, Page404 };
